Fixed python module for docker

+++ utilities/docker/centos/gridlabd.sh
@@ -73,13 +73,29 @@ else
 	echo "WARNING: ${ARMA} not found -- armadillo will not be included in this build" > /dev/stderr
 fi
 
+# cleanup third-party sources
+echo "Cleaning up third-party sources"
+rm -rf ._armadillo-7.800.1 armadillo-7.800.1.tar armadillo-7.800.1/ mysql-connector-c-6.1.11-linux-glibc2.12-x86_64.tar mysql-connector-c-6.1.11-linux-glibc2.12-x86_64/ xerces-c-src_2_8_0.tar
+git checkout .
+
+# debugger support
+if [ "${ENABLE_GDB}" == "yes" ]; then
+	echo "Added support for gdb"
+	debuginfo-install glibc-2.17-222.el7.x86_64 libgcc-4.8.5-28.el7_5.1.x86_64 libstdc++-4.8.5-28.el7_5.1.x86_64
+	CFLAGS='-w -O0 -g'
+	CXXFLAGS='-w -O0 -g'
+else
+	CFLAGS='-w -O3'
+	CXXFLAGS='-w -O3'
+fi
+
 # install gridlabd
 cd /usr/local/src/gridlabd
 if [ -f customize -a ! -z "${ENABLE}" ]; then
 	./customize enable ${ENABLE} || true
 fi
 autoreconf -isf
-./configure --enable-silent-rules --prefix=/usr/local --with-mysql=$MYSQLOPT 'CXXFLAGS=-w -O3' 'CFLAGS=-w -O3'
+./configure --enable-silent-rules --prefix=/usr/local --with-mysql=$MYSQLOPT
 make install
 if [ "$BRANCH" = "master" ]
 then 
@@ -108,8 +124,10 @@ mkdir -p /usr/local/var
 adduser -d /usr/local/var/gridlabd gridlabd
 
 # clean up
-echo "Cleaning up cloned repositories"
-rm -rf /usr/local/src/*
+if [ "${KEEP_SOURCE:-no}" == "yes" ]; then
+	echo "Cleaning up source code"
+	rm -rf /usr/local/src/*
+fi
 
 # done
 echo "Build of $REPO/$BRANCH done"
